+++ packages/picasso/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.77.2](https://github.com/toptal/picasso/compare/@toptal/picasso@4.77.1...@toptal/picasso@4.77.2) (2020-07-14)
+
+
+### Bug Fixes
+
+* ignore titleCase for Autocomplete options ([#1432](https://github.com/toptal/picasso/issues/1432)) ([e35899e](https://github.com/toptal/picasso/commit/e35899e1ecc1148f2b049d24594574af6005125a))
+
+
+
+
+
+## [4.77.1](https://github.com/toptal/picasso/compare/@toptal/picasso@4.77.0...@toptal/picasso@4.77.1) (2020-07-14)
+
+
+### Bug Fixes
+
+* ignore titleCase for Select options ([#1431](https://github.com/toptal/picasso/issues/1431)) ([5b24f3e](https://github.com/toptal/picasso/commit/5b24f3e7c4f5ccfff5400de7d233dedebcaa2002))
+
+
+
+
+
+# [4.77.0](https://github.com/toptal/picasso/compare/@toptal/picasso@4.76.1...@toptal/picasso@4.77.0) (2020-07-14)
+
+
+### Features
+
+* add dispute and pause icons ([#1427](https://github.com/toptal/picasso/issues/1427)) ([c06cb46](https://github.com/toptal/picasso/commit/c06cb4645f1205a44e8b370b93664a977bc5ea74))
+
+
+
+
+
 ## [4.76.1](https://github.com/toptal/picasso/compare/@toptal/picasso@4.76.0...@toptal/picasso@4.76.1) (2020-07-13)